Change -v to -n for better if statement checks
if [[ -n $VIRTUAL_ENV ]] - checks if the VIRTUAL_ENV variable is non-empty.
The -n operator returns true if the length of the string is non-zero.

if [[ -v $VIRTUAL_ENV ]] - checks if the VIRTUAL_ENV variable has been set, regardless of whether it is empty or not.
The -v operator returns true if the variable has been assigned a value, even if that value is an empty string.